Move cli args to runner
